extract unbounded base components module from base components
