+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/util/transformer/UnionTransformersTestModeSpec.scala
@@ -11,6 +11,7 @@ import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.build.{GraphBuilder, ScenarioBuilder}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
+import pl.touk.nussknacker.engine.flink.FlinkBaseUnboundedComponentProvider
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.flink.util.source.CollectionSource
 import pl.touk.nussknacker.engine.graph.node
@@ -119,7 +120,7 @@ class UnionTransformersTestModeSpec
     )
     LocalModelData(
       ConfigFactory.empty(),
-      sourceComponent :: FlinkBaseComponentProvider.Components,
+      sourceComponent :: FlinkBaseUnboundedComponentProvider.Components ::: FlinkBaseComponentProvider.Components,
       configCreator = new ConfigCreatorWithCollectingListener(collectingListener),
     )
   }
