+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/AggregatesSpec.scala
@@ -4,19 +4,7 @@ import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedObjectTypingResult, TypingResult, Unknown}
-import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.aggregates.{
-  AverageAggregator,
-  CountWhenAggregator,
-  FirstAggregator,
-  LastAggregator,
-  ListAggregator,
-  MapAggregator,
-  MaxAggregator,
-  MinAggregator,
-  OptionAggregator,
-  SetAggregator,
-  SumAggregator
-}
+import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.aggregates._
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 
 import java.lang.{Integer => JInt, Long => JLong}
