+++ engine/flink/components/base-unbounded/src/main/resources/META-INF/services/pl.touk.nussknacker.engine.api.component.ComponentProvider
@@ -1,0 +1,1 @@
+pl.touk.nussknacker.engine.flink.FlinkBaseUnboundedComponentProvider
