+++ engine/flink/components/base/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/FlinkBaseComponentProvider.scala
@@ -5,13 +5,6 @@ import pl.touk.nussknacker.engine.api.component.{ComponentDefinition, ComponentP
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, SinkFactory}
 import pl.touk.nussknacker.engine.common.components.DecisionTable
 import pl.touk.nussknacker.engine.flink.util.sink.EmptySink
-import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.AggregateWindowsConfig
-import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.sampleTransformers.{
-  SessionWindowAggregateTransformer,
-  SlidingAggregateTransformerV2,
-  TumblingAggregateTransformer
-}
-import pl.touk.nussknacker.engine.flink.util.transformer.join.{FullOuterJoinTransformer, SingleSideJoinTransformer}
 import pl.touk.nussknacker.engine.util.config.DocsConfig
 
 class FlinkBaseComponentProvider extends ComponentProvider {
@@ -20,9 +13,8 @@ class FlinkBaseComponentProvider extends ComponentProvider {
   override def resolveConfigForExecution(config: Config): Config = config
 
   override def create(config: Config, dependencies: ProcessObjectDependencies): List[ComponentDefinition] = {
-    val docsConfig             = DocsConfig(config)
-    val aggregateWindowsConfig = AggregateWindowsConfig.loadOrDefault(config)
-    FlinkBaseComponentProvider.create(docsConfig, aggregateWindowsConfig)
+    val docsConfig = DocsConfig(config)
+    FlinkBaseComponentProvider.create(docsConfig)
   }
 
   override def isCompatible(version: NussknackerVersion): Boolean = true
@@ -33,46 +25,20 @@ class FlinkBaseComponentProvider extends ComponentProvider {
 object FlinkBaseComponentProvider {
 
   val Components: List[ComponentDefinition] =
-    create(DocsConfig.Default, AggregateWindowsConfig.Default)
+    create(DocsConfig.Default)
 
-  def create(docsConfig: DocsConfig, aggregateWindowsConfig: AggregateWindowsConfig): List[ComponentDefinition] = {
+  def create(docsConfig: DocsConfig): List[ComponentDefinition] = {
     import docsConfig._
 
-    // When adding/changing stateful components, corresponding changes should be done in LiteBaseComponentProvider!
-    val statelessComponents = List(
+    List(
       ComponentDefinition("for-each", ForEachTransformer).withRelativeDocs("BasicNodes#foreach"),
       ComponentDefinition("union", UnionTransformer).withRelativeDocs("BasicNodes#union"),
       ComponentDefinition("dead-end", SinkFactory.noParam(EmptySink)).withRelativeDocs("DataSourcesAndSinks#deadend"),
-      ComponentDefinition("periodic", PeriodicSourceFactory).withRelativeDocs("DataSourcesAndSinks#periodic"),
       ComponentDefinition(
         name = "decision-table",
         component = DecisionTable
       ).withRelativeDocs("BasicNodes#decisiontable")
     )
-
-    val statefulComponents = List(
-      ComponentDefinition("union-memo", UnionWithMemoTransformer).withRelativeDocs("DataSourcesAndSinks#unionmemo"),
-      ComponentDefinition("previousValue", PreviousValueTransformer).withRelativeDocs(
-        "DataSourcesAndSinks#previousvalue"
-      ),
-      ComponentDefinition("aggregate-sliding", SlidingAggregateTransformerV2).withRelativeDocs(
-        "AggregatesInTimeWindows#sliding-window"
-      ),
-      ComponentDefinition("aggregate-tumbling", new TumblingAggregateTransformer(aggregateWindowsConfig))
-        .withRelativeDocs("AggregatesInTimeWindows#tumbling-window"),
-      ComponentDefinition("aggregate-session", SessionWindowAggregateTransformer).withRelativeDocs(
-        "AggregatesInTimeWindows#session-window"
-      ),
-      ComponentDefinition("single-side-join", SingleSideJoinTransformer).withRelativeDocs(
-        "AggregatesInTimeWindows#single-side-join"
-      ),
-      ComponentDefinition("full-outer-join", FullOuterJoinTransformer).withRelativeDocs(
-        "AggregatesInTimeWindows#full-outer-join"
-      ),
-      ComponentDefinition("delay", DelayTransformer).withRelativeDocs("DataSourcesAndSinks#delay")
-    )
-
-    statefulComponents ++ statelessComponents
   }
 
 }
