+++ nussknacker-dist/src/universal/conf/dev-application.conf
@@ -99,7 +99,7 @@ scenarioTypes {
   "streaming-dev": {
     deploymentConfig: ${flinkDeploymentConfig}
     modelConfig: {
-      classPath: ["components/flink-dev/flinkTable.jar", "model/devModel.jar", "model/flinkExecutor.jar", "components/flink", "flink-dropwizard-metrics-deps/"]
+      classPath: ["model/devModel.jar", "model/flinkExecutor.jar", "components/flink", "flink-dropwizard-metrics-deps/"]
       rocksDB: {
         enable: true
         enable: ${?FLINK_ROCKSDB_ENABLE}
@@ -113,12 +113,6 @@ scenarioTypes {
           }
           docsUrl: "https://nussknacker.io"
         }
-      }
-      components.flinkTable {
-        // The table configuration file should be mounted on the Flink Taskmanager container under this path. You can
-        // override this path through the 'SQL_DATASOURCES_FILE_PATH' env variable for a given Flink service.
-        tableDefinitionFilePath: "../../../nussknacker-dist/src/universal/conf/tables-definition.sql"
-        tableDefinitionFilePath: ${?SQL_DATASOURCES_FILE_PATH}
       }
     }
     category: "Category1"
@@ -143,6 +137,23 @@ scenarioTypes {
       }
     }
     category: "Category2"
+  }
+  "batch-dev": {
+    deploymentConfig: ${flinkDeploymentConfig}
+    modelConfig: {
+      classPath: ["components/flink-dev/flinkTable.jar", "model/devModel.jar", "model/flinkExecutor.jar", "components/flink/flinkBase.jar", "flink-dropwizard-metrics-deps/"]
+      rocksDB: {
+        enable: true
+        enable: ${?FLINK_ROCKSDB_ENABLE}
+      }
+      components.flinkTable {
+        // The table configuration file should be mounted on the Flink Taskmanager container under this path. You can
+        // override this path through the 'SQL_DATASOURCES_FILE_PATH' env variable for a given Flink service.
+        tableDefinitionFilePath: "../../../nussknacker-dist/src/universal/conf/tables-definition.sql"
+        tableDefinitionFilePath: ${?SQL_DATASOURCES_FILE_PATH}
+      }
+    }
+    category: "Batch"
   }
   "periodic-dev": {
     deploymentConfig: {
